Remove confusion for 'Add an issue', fixes #44
Changed wording on search page to make it clear that when a user doesn't
find what they are looking for on the search page, they are not adding a
issue against Bountysource's search algorithm, but rather adding a
missing issue.

Added validation messages to 'add issue' page to inform user when they
have bad input. Submit button should always be enabled, as users don't
readily understand that is it disabled because of invalid input.